Added healing spells functionality

+++ 60-377/Assignment 4/Classes/Unit.cs
@@ -12,7 +12,7 @@ public abstract class Unit : MonoBehaviour
         this.health -= damage;
     }
 
-    public void RegenerateHealth(int health)
+    public void RegenerateHealth(float health)
     {
         this.health += health;
     }
